+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## [Unreleased]
 
+ - Updated Facebook Graph API version to v10.0
  - Updated RTBHouse API version to 5
- - Updated Facebook API version to 8
  - Allow overwriting configuration values from cli
 
 ## [1.3.8] - 2020-02-04